[YUNIKORN-3061] Refactor recovery_and_restart E2E test suite structure
